requirements.txt
django
Shopifyapi
python-dotenv
psycopg2
aiohttp
pytz
django-timezone-field
djangorestframework
requests
whitenoise
gunicorn
psycopg2
django-cors-headers
djangorestframework-simplejwt
python-dateutil
pymongo
celery 
redis
boto3
django-storages
django-celery-beat
